# Authorized students: Optimize query for lower-access users