--- a/app/models/educator_label.rb
+++ b/app/models/educator_label.rb
@@ -67,8 +67,7 @@ class EducatorLabel < ApplicationRecord
     end
 
     # Levels link: Anyone at SHS (can also be added manually)
-    shs_school_id = School.find_by_local_id('SHS').try(:id)
-    if shs_school_id.present? && educator.school_id == shs_school_id
+    if educator.school.present? && educator.school.local_id == 'SHS'
       dynamic_labels << 'should_show_levels_shs_link'
     end